refactor: 클래스 리펙토링
- 필요없는 클래스 제거
- 패키지 구조 변경
