+++ src/main/kotlin/nextstep/mission/lotto/vo/WinningResult.kt
@@ -1,7 +1,5 @@
 package nextstep.mission.lotto.vo
 
-import nextstep.mission.lotto.WinningPrize
-
 @JvmInline
 value class WinningResult(private val result: Map<WinningPrize, Int>) {
